Decouple finding lcs from diff_sequence function
This allows us to swap a faster algorithm in more easily, also it stops
us holding onto the big matrix for too long, we now reduce this to the
lcs list as early as possible.

creating diff blocks now traverses the lcs list forwards which is a
little easier to understand. It has had the added benefit of making the
code more comprehensible too!