add more tests for patch, fix named tuple patching.
namedtuples are frankly a bit of a pain the arse. They don't conform to
the rule of normal sequence construction wherby you pass in an iterable.
Instead you either pass in items as args or you use the '_make'
attribute. Mind you the differences between str and other sequences is
also a little irritating too.

It would be nice if all sequences followed the same protocol for
construction...