added ordered mapping patch tests and fixed patch ordered mapping.
patch ordered mapping is fixed in the sense that the tests already
created for it now pass. However more testing is required for all of the
patch functions, as the desired behaviour is not completely captured in the
tests yet. There are certainly some obscure edge cases which will be
worth checking.