improved error messages and added more patch tests. 100% code coverage
I already know that OrderedDict has a bug from code insepction. It adds
the unchanged items from the diff rather than from the patch target. It
needs a rewrite that does better validation. Probably going to have to
treat it a lot more like a sequence. Need to write tests and fix it.